+++ getVersionTags.sh
@@ -1,13 +1,13 @@
 #!/bin/sh
 
 # get the latest version
-pyincore=$(curl -s 'https://api.github.com/repos/IN-CORE/pyincore/tags' | jq '.[0].name')
-pyincoreViz=$(curl -s 'https://api.github.com/repos/IN-CORE/pyincore-viz/tags' | jq '.[0].name')
-pyincoreData=$(curl -s 'https://api.github.com/repos/IN-CORE/pyincore-data/tags' | jq '.[0].name')
-incoreUI=$(curl -s 'https://api.github.com/repos/IN-CORE/incore-ui/tags' | jq '.[0].name')
-incoreServices=$(curl -s 'https://api.github.com/repos/IN-CORE/incore-services/tags' | jq '.[0].name')
-incoreLab=$(curl -s 'https://api.github.com/repos/IN-CORE/incore-lab/tags' | jq '.[0].name')
-incoreDocs=$(curl -s 'https://api.github.com/repos/IN-CORE/incore-docs/tags' | jq '.[0].name')
+pyincore=$(curl -s 'https://api.github.com/repos/IN-CORE/pyincore/releases/latest' | jq '.tag_name')
+pyincoreViz=$(curl -s 'https://api.github.com/repos/IN-CORE/pyincore-viz/releases/latest' | jq '.tag_name')
+pyincoreData=$(curl -s 'https://api.github.com/repos/IN-CORE/pyincore-data/releases/latest' | jq '.tag_name')
+incoreUI=$(curl -s 'https://api.github.com/repos/IN-CORE/incore-ui/releases/latest' | jq '.tag_name')
+incoreServices=$(curl -s 'https://api.github.com/repos/IN-CORE/incore-services/releases/latest' | jq '.tag_name')
+incoreLab=$(curl -s 'https://api.github.com/repos/IN-CORE/incore-lab/releases/latest' | jq '.tag_name')
+incoreDocs=$(curl -s 'https://api.github.com/repos/IN-CORE/incore-docs/releases/latest' | jq '.tag_name')
 
 if [ -z "$pyincore" ] || [ -z "$pyincoreViz" ] || [ -z "$incoreUI" ] || [ -z "$incoreServices" ] || [ -z "$incoreLab" ] || [ -z "$incoreDocs" ]
 then