+++ src/components/Profile.jsx
@@ -109,7 +109,7 @@ const useStyles = makeStyles({
 });
 
 export default function Profile(props) {
-	const {loginError, datasetUsage, hazardUsage, allocations} = props;
+	const {loginError, usage, getUsage, allocations, getAllocations} = props;
 	const [authError, setAuthError] = useState(false);
 	const [dataEntityPie, setDataEntityPie] = useState({});
 	const [dataFileSizePie, setDataFileSizePie] = useState({});
@@ -127,6 +127,10 @@ export default function Profile(props) {
 		// logged in
 		if (config.hostname.includes("localhost") || (authorization !== undefined && authorization !== "" && authorization !== null)) {
 			setAuthError(false);
+
+			// need to refresh from endpoint when page load; just in case people deleted stuff on other viewers
+			getAllocations();
+			getUsage();
 		}
 		// not logged in
 		else {
@@ -135,22 +139,17 @@ export default function Profile(props) {
 	}, []);
 
 	useEffect(() => {
-		if (allocations !== undefined && Object.keys(allocations).length > 0
-			&& datasetUsage !== undefined && Object.keys(datasetUsage).length > 0){
-			let pieChartConfig = configurePieCharts(allocations, datasetUsage, "datasetUsage");
-			setDataEntityPie(pieChartConfig["entity"]);
-			setDataFileSizePie(pieChartConfig["fileSize"]);
-		}
-	}, [datasetUsage, allocations]);
+		if (allocations !== undefined && Object.keys(allocations).length > 0 &&
+			usage !== undefined && Object.keys(usage).length > 0){
+			let dataPieChartConfig = configurePieCharts(allocations, usage, "datasetUsage");
+			setDataEntityPie(dataPieChartConfig["entity"]);
+			setDataFileSizePie(dataPieChartConfig["fileSize"]);
 
-	useEffect(() => {
-		if (allocations !== undefined && Object.keys(allocations).length > 0
-			&& hazardUsage !== undefined && Object.keys(hazardUsage).length > 0){
-			let pieChartConfig = configurePieCharts(allocations, hazardUsage, "hazardUsage");
-			setHazardEntityPie(pieChartConfig["entity"]);
-			setHazardFileSizePie(pieChartConfig["fileSize"]);
+			let hazardPieChartConfig = configurePieCharts(allocations, usage, "hazardUsage");
+			setHazardEntityPie(hazardPieChartConfig["entity"]);
+			setHazardFileSizePie(hazardPieChartConfig["fileSize"]);
 		}
-	}, [hazardUsage, allocations]);
+	}, [usage, allocations]);
 
 	// for any auth error
 	useEffect(() =>{
@@ -166,44 +165,57 @@ export default function Profile(props) {
 		// to create a deep copy !important
 		let defaultEntityPieConfig = JSON.parse(JSON.stringify(chartConfig.pieChartConfig));
 		let defaultFileSizePieConfig = JSON.parse(JSON.stringify(chartConfig.pieChartConfig));
+		let totalNum = -999;
+		let totalFileSizeByte = -999;
+		let totalFileSize = "NA";
+		let totalNumAllocated = -999;
+		let totalBytesAllocated = -999;
+		let totalBytesTextAllocated = "NA";
 
-		const totalNumAllocated = type === "datasetUsage" ?
-			allocations["total_number_of_datasets"]
-			:
-			allocations["total_number_of_hazards"];
-		const totalBytesAllocated = type === "datasetUsage"?
-			allocations["total_file_size_of_datasets_byte"]
-			:
-			allocations["total_file_size_of_hazard_datasets_byte"];
-		const totalBytesTextAllocated = type === "datasetUsage" ?
-			allocations["total_file_size_of_datasets"]
-			:
-			allocations["total_file_size_of_hazard_datasets"];
+		if (type === "datasetUsage"){
+			totalNum = usage["total_number_of_datasets"];
+			totalFileSizeByte = usage["total_file_size_of_datasets_byte"];
+			totalFileSize = usage["total_file_size_of_datasets"];
+			totalNumAllocated = allocations["total_number_of_datasets"];
+			totalBytesAllocated = allocations["total_file_size_of_datasets_byte"];
+			totalBytesTextAllocated = allocations["total_file_size_of_datasets"];
+		}
+		else if (type === "hazardUsage"){
+			totalNum = usage["total_number_of_hazards"];
+			totalFileSizeByte = usage["total_file_size_of_hazard_datasets_byte"];
+			totalFileSize = usage["total_file_size_of_hazard_datasets"];
+			totalNumAllocated = allocations["total_number_of_hazards"];
+			totalBytesAllocated = allocations["total_file_size_of_hazard_datasets_byte"];
+			totalBytesTextAllocated = allocations["total_file_size_of_hazard_datasets"];
+		}
+		else{
+			console.log(`${type} not supported!`);
+		}
 
 		defaultEntityPieConfig["series"][0]["data"] = [
 			{
 				name: "Available",
-				y: totalNumAllocated - usage["total_number_of_datasets"]
+				y: totalNumAllocated - totalNum
 			},
 			{
 				name: "Used",
-				y: usage["total_number_of_datasets"]
-			}
-		];
+				y: totalNum
+			}];
 		defaultEntityPieConfig["title"]["text"] = "Entities";
-		defaultEntityPieConfig["subtitle"]["text"] = `Used ${usage["total_number_of_datasets"]} of ${totalNumAllocated}`;
+		defaultEntityPieConfig["subtitle"]["text"] = `Used ${totalNum} of ${totalNumAllocated}`;
+
 
 		defaultFileSizePieConfig["series"][0]["data"] = [
 			{
 				name: "Available",
-				y: totalBytesAllocated - usage["total_file_size_byte"]
+				y: totalBytesAllocated - totalFileSizeByte
 			},
 			{
 				name: "Used",
-				y: usage["total_file_size_byte"]
+				y: totalFileSizeByte
 			}];
 		defaultFileSizePieConfig["title"]["text"] = "File Size";
-		defaultFileSizePieConfig["subtitle"]["text"] = `Used ${usage["total_file_size"]} of ${totalBytesTextAllocated}`;
+		defaultFileSizePieConfig["subtitle"]["text"] = `Used ${totalFileSize} of ${totalBytesTextAllocated}`;
 
 		return {"entity": defaultEntityPieConfig, "fileSize": defaultFileSizePieConfig};
 	};